[Optimization] Convert to type-generic math

+++ AsyncDisplayKit/ASEditableTextNode.mm
@@ -11,6 +11,7 @@
 #import "ASEditableTextNode.h"
 
 #import <objc/message.h>
+#import <tgmath.h>
 
 #import "ASDisplayNode+Subclasses.h"
 #import "ASEqualityHelpers.h"
@@ -238,9 +239,9 @@
 {
   ASTextKitComponents *displayedComponents = [self isDisplayingPlaceholder] ? _placeholderTextKitComponents : _textKitComponents;
   CGSize textSize = [displayedComponents sizeForConstrainedWidth:constrainedSize.width];
-  CGFloat width = ceilf(textSize.width + _textContainerInset.left + _textContainerInset.right);
-  CGFloat height = ceilf(textSize.height + _textContainerInset.top + _textContainerInset.bottom);
-  return CGSizeMake(fminf(width, constrainedSize.width), fminf(height, constrainedSize.height));
+  CGFloat width = ceil(textSize.width + _textContainerInset.left + _textContainerInset.right);
+  CGFloat height = ceil(textSize.height + _textContainerInset.top + _textContainerInset.bottom);
+  return CGSizeMake(fmin(width, constrainedSize.width), fmin(height, constrainedSize.height));
 }
 
 - (void)layout
