+++ AsyncDisplayKit/Details/ASHighlightOverlayLayer.mm
@@ -11,6 +11,7 @@
 #import "ASHighlightOverlayLayer.h"
 
 #import <UIKit/UIKit.h>
+#import <tgmath.h>
 
 #import "ASInternalHelpers.h"
 
@@ -86,7 +87,7 @@ static const UIEdgeInsets padding = {2, 4, 1.5, 4};
     if (targetLayer != nil) {
       rect = [self convertRect:rect fromLayer:targetLayer];
     }
-    rect = CGRectMake(roundf(rect.origin.x), roundf(rect.origin.y), roundf(rect.size.width), roundf(rect.size.height));
+    rect = CGRectMake(round(rect.origin.x), round(rect.origin.y), round(rect.size.width), round(rect.size.height));
 
     CGFloat minX = rect.origin.x - padding.left;
     CGFloat maxX = CGRectGetMaxX(rect) + padding.right;
